Refactor & remove extra
