+++ src/views/trip-event-list-wrapper.js
